[scripting] "fixed" exception in js console
The current solution is an unsatisfactory workaround. For some reason, the
js context has stopped throwing exceptions back at us, it just returns Strings.
That's so bad it almost can't be right... but I can't seem to get it to
work in any other way either. Given all this, I take the String returned
when something goes wrong, and replace it with a slightly nicer error message.

git-svn-id: https://bioclipse.svn.sourceforge.net/svnroot/bioclipse/bioclipse2/trunk@9102 fcb5ba71-d80d-0410-8237-ba3920747fcc