Added keylistener to jmolPanel to lisen for CTRL/COMMAND-W
This adds a key listener that listens for W with
modifier CTRL or COMMAND depending if it is mac 
or not, if it is close editor is called.

This is kinda a qick fix dunno if there is a real fix
like getting those events out to eclpse.
Fixes bug #835

git-svn-id: https://bioclipse.svn.sourceforge.net/svnroot/bioclipse/bioclipse2/trunk@10207 fcb5ba71-d80d-0410-8237-ba3920747fcc