[GroovyConsoleView] made printing work again
The way scripting consoles work internally had changed since the Groovy
console was buried. Specifically, the executecommand() method is now
itself responsible for printing the result back to the console. This change
was necessary to make async evaluation work. The Groovy console doesn't
make use of async, but it still needs to conform.