Added silentAfter timeout to @SilentNotification
Now the SilentNotification will start after a certain "silentAfter" amount of time has passed. This value can be set using the annotation and the default value is 500ms. That is if the job is finnished within half a second after it is started it will opo up the result at once and if it takes longer the user will have to click in the progress view to see the result.