Fixed version number on splash screen
Moved the version number to the top left of the Bioclipse splash and set it to being the same version number as the plugin net.bioclipse.ui has. For the moment it is 2.4.0.20100709 (expected 2.4.1....) but release managers should make sure the net.bioclipse.ui plugin has the version number Bioclipse should have.