warn about UndeclaredThrowableException in log
If a manager method throws an Exception that its corresponding interface does not declare this Exception will be thrown. Now a warning about it will be printed to the log making it easier to debug. The main reason this can be a problem is because the BioclipseException is a checked exception (which keeps on biting us). Now if a manager interface author forgets to write "throws BioclipseException" where needed at least a warning is written to the log when the code is run.